// src/app/model/usuarioPiDto.ts
export class UsuarioPiDto {
    codUsuario: number;
    codTipoUsuario : number;
    tipoUsuario: string;
    codTipoAutenticacao : number;
    tipoAutenticacao: string;
    codRole:number;
    role: string;
    codExterno: number;
    nome: string;
    cpf: string;
    login: string;
    email:string;
    codAtivo : boolean;
    ativo: string;
}